+++ src/pages/PostsManagerPage.tsx
@@ -8,8 +8,6 @@ import {
   Trash2,
 } from "lucide-react"
 import { useState } from "react"
-import { commentApi } from "../entities/comment/api/commentApi"
-import { Comment, NewComment } from "../entities/comment/model/types"
 import { SortOrder, usePostQueryParams, usePostsQuery } from "../entities/post"
 import { usePostsQueryProps } from "../entities/post/api/usePostsQuery"
 import { usePostTagsQuery } from "../entities/post/api/usePostTagsQuery"
@@ -46,6 +44,7 @@ import {
   Textarea,
   TextHighlighter,
 } from "../shared/ui"
+import { PostDetailDialog } from "../widgets/post"
 
 const PostsManager = () => {
   const {
@@ -66,16 +65,6 @@ const PostsManager = () => {
   const [showAddDialog, setShowAddDialog] = useState(false)
   const [showEditDialog, setShowEditDialog] = useState(false)
 
-  const [comments, setComments] = useState<Record<string, Comment[]>>({})
-  const [selectedComment, setSelectedComment] = useState<Comment | null>(null)
-  const [newComment, setNewComment] = useState<NewComment>({
-    body: "",
-    postId: null,
-    userId: 1,
-  })
-
-  const [showAddCommentDialog, setShowAddCommentDialog] = useState(false)
-  const [showEditCommentDialog, setShowEditCommentDialog] = useState(false)
   const [showPostDetailDialog, setShowPostDetailDialog] = useState(false)
 
   const [showUserModal, setShowUserModal] = useState(false)
@@ -118,78 +107,9 @@ const PostsManager = () => {
     })
   }
 
-  // 댓글 가져오기
-  const fetchComments = async (postId: number) => {
-    if (comments[postId]) return // 이미 불러온 댓글이 있으면 다시 불러오지 않음
-
-    const data = await commentApi.fetchComments(postId)
-    if (data) {
-      setComments((prev) => ({ ...prev, [postId]: data.comments }))
-    }
-  }
-
-  // 댓글 추가
-  const addComment = async () => {
-    const data = await commentApi.addComment(newComment)
-    if (data) {
-      setComments((prev) => ({
-        ...prev,
-        [data.postId]: [...(prev[data.postId] || []), data],
-      }))
-      setShowAddCommentDialog(false)
-      setNewComment({ body: "", postId: null, userId: 1 })
-    }
-  }
-
-  // 댓글 업데이트
-  const updateComment = async () => {
-    if (!selectedComment) return
-
-    const data = await commentApi.updateComment(selectedComment)
-    if (data) {
-      setComments((prev) => ({
-        ...prev,
-        [data.postId]: prev[data.postId].map((comment) =>
-          comment.id === data.id ? data : comment,
-        ),
-      }))
-      setShowEditCommentDialog(false)
-    }
-  }
-
-  // 댓글 삭제
-  const deleteComment = async (
-    id: Comment["id"],
-    postId: Comment["postId"],
-  ) => {
-    await commentApi.deleteComment(id)
-    setComments((prev) => ({
-      ...prev,
-      [postId]: prev[postId].filter((comment) => comment.id !== id),
-    }))
-  }
-
-  // 댓글 좋아요
-  const likeComment = async (id: Comment["id"], postId: Comment["postId"]) => {
-    const data = await commentApi.likeComment(
-      id,
-      comments[postId].find((c) => c.id === id)?.likes,
-    )
-
-    if (data) {
-      setComments((prev) => ({
-        ...prev,
-        [postId]: prev[postId].map((comment) =>
-          comment.id === data.id ? data : comment,
-        ),
-      }))
-    }
-  }
-
   // 게시물 상세 보기
   const openPostDetail = (post: Post) => {
     setSelectedPost(post)
-    fetchComments(post.id)
     setShowPostDetailDialog(true)
   }
 
@@ -298,69 +218,6 @@ const PostsManager = () => {
     </Table>
   )
 
-  // 댓글 렌더링
-  const renderComments = (postId: Post["id"]) => (
-    <div className="mt-2">
-      <div className="flex items-center justify-between mb-2">
-        <h3 className="text-sm font-semibold">댓글</h3>
-        <Button
-          size="sm"
-          onClick={() => {
-            setNewComment((prev) => ({ ...prev, postId }))
-            setShowAddCommentDialog(true)
-          }}
-        >
-          <Plus className="w-3 h-3 mr-1" />
-          댓글 추가
-        </Button>
-      </div>
-      <div className="space-y-1">
-        {comments[postId]?.map((comment) => (
-          <div
-            key={comment.id}
-            className="flex items-center justify-between text-sm border-b pb-1"
-          >
-            <div className="flex items-center space-x-2 overflow-hidden">
-              <span className="font-medium truncate">
-                {comment.user.username}:
-              </span>
-              <span className="truncate">
-                <TextHighlighter text={comment.body} highlight={searchQuery} />
-              </span>
-            </div>
-            <div className="flex items-center space-x-1">
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => likeComment(comment.id, postId)}
-              >
-                <ThumbsUp className="w-3 h-3" />
-                <span className="ml-1 text-xs">{comment.likes}</span>
-              </Button>
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => {
-                  setSelectedComment(comment)
-                  setShowEditCommentDialog(true)
-                }}
-              >
-                <Edit2 className="w-3 h-3" />
-              </Button>
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => deleteComment(comment.id, postId)}
-              >
-                <Trash2 className="w-3 h-3" />
-              </Button>
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  )
-
   return (
     <Card className="w-full max-w-6xl mx-auto">
       <CardHeader>
@@ -395,7 +252,6 @@ const PostsManager = () => {
               value={selectedTag}
               onValueChange={(value) => {
                 updateQueryParam({ tag: value })
-                // fetchPostsByTag(value)
               }}
             >
               <SelectTrigger className="w-[180px]">
@@ -554,80 +410,13 @@ const PostsManager = () => {
         </DialogContent>
       </Dialog>
 
-      {/* 댓글 추가 대화상자 */}
-      <Dialog
-        open={showAddCommentDialog}
-        onOpenChange={setShowAddCommentDialog}
-      >
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>새 댓글 추가</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <Textarea
-              placeholder="댓글 내용"
-              value={newComment.body}
-              onChange={(e) =>
-                setNewComment({ ...newComment, body: e.target.value })
-              }
-            />
-            <Button onClick={addComment}>댓글 추가</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
-
-      {/* 댓글 수정 대화상자 */}
-      <Dialog
-        open={showEditCommentDialog}
-        onOpenChange={setShowEditCommentDialog}
-      >
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>댓글 수정</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <Textarea
-              placeholder="댓글 내용"
-              value={selectedComment?.body || ""}
-              onChange={(e) => {
-                if (selectedComment) {
-                  setSelectedComment({
-                    ...selectedComment,
-                    body: e.target.value,
-                  })
-                }
-              }}
-            />
-            <Button onClick={updateComment}>댓글 업데이트</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
-
       {/* 게시물 상세 보기 대화상자 */}
-      <Dialog
+      <PostDetailDialog
         open={showPostDetailDialog}
         onOpenChange={setShowPostDetailDialog}
-      >
-        <DialogContent className="max-w-3xl">
-          <DialogHeader>
-            <DialogTitle>
-              <TextHighlighter
-                text={selectedPost?.title}
-                highlight={searchQuery}
-              />
-            </DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <p>
-              <TextHighlighter
-                text={selectedPost?.body}
-                highlight={searchQuery}
-              />
-            </p>
-            {selectedPost && renderComments(selectedPost?.id)}
-          </div>
-        </DialogContent>
-      </Dialog>
+        searchQuery={searchQuery}
+        selectedPost={selectedPost}
+      />
 
       {/* 사용자 모달 */}
       {selectedUser && (
